sherpa.image now imports but ds9 connection fails
The ds9_backend module failed to import because of an un-needed import of
izip. Several PEP8 and minor style changes have been made.

The test failure when ds9 is installed is not very useful, since it is
the "standard" Sherpa DS9Err('noimage') message and so hides the actual
cause of the problem. From some manual testing it looks likely related
to bytes and/or bytearrays.

I have asked before, but does it make sense to keep this module when
there is the pyds9 module (although I don't know if that supports python3
at this time)?